+++ python/setup.py
@@ -1,0 +1,14 @@
+''' setup for thinwalltopography '''
+import setuptools
+
+setuptools.setup(
+    name="thinwalltopography",
+    version="0.0.1",
+    author="",
+    author_email="",
+    description=(""),
+    license="",
+    keywords="",
+    url="",
+    packages=['thinwalltopography']
+)
