+++ python/thinwalltopography/ice9.py
