+++ python/thinwalltopography/merge_topog_tiles.py
@@ -2,7 +2,7 @@
 
 import sys, getopt
 import datetime, os, subprocess
-import GMesh
+from thinwalltopography import GMesh
 import imp
 import netCDF4
 import numpy as np
@@ -50,14 +50,14 @@ def write_topog(h,hstd,hmin,hmax,hx,hy,fnam=None,format='NETCDF3_CLASSIC',descri
     h_y=fout.createVariable('y','f8',('ny','nx'))
     h_y.units='degrees'
     h_y[:]=hy
-#    string=fout.createDimension('string',255)    
+#    string=fout.createDimension('string',255)
 #    tile=fout.createVariable('tile','S1',('string'))
 #    tile[0]='t'
 #    tile[1]='i'
 #    tile[2]='l'
 #    tile[3]='e'
 #    tile[4]='1'
-    
+
     #global attributes
     if(not no_changing_meta):
     	fout.history = history
@@ -103,7 +103,7 @@ def main(argv):
         print(err)
         usage(scriptbasename)
         sys.exit(2)
-        
+
     for opt, arg in opts:
         if opt == '-h':
             usage()
@@ -131,10 +131,10 @@ def main(argv):
         hist = hist + " on "+ str(datetime.date.today()) + " on platform "+ host
 
     desc = "This is a model topography file generated by the refine-sampling method from source topography. "
-    
+
     source =""
     if(not no_changing_meta):
-        source =  source + scriptpath + " had git hash " + scriptgithash + scriptgitMod 
+        source =  source + scriptpath + " had git hash " + scriptgithash + scriptgitMod
         source =  source + ". To obtain the grid generating code do: git clone  https://github.com/nikizadehgfdl/thin-wall-topography.git ; cd thin-wall-topography;  git checkout "+scriptgithash
 
     tiles = tilefiles.split(',')
@@ -159,29 +159,29 @@ def main(argv):
     print(" h.shape=",height.shape,"  h.first=",height[0,0]," h.last=",height[-1,0])
     for h in heights[1:] :
         print(" h.shape=",h.shape," h.first=",h[0,0]," h.last=",h[-1,0])
-        height = np.concatenate((height[:-1,:],h),axis=0) 
+        height = np.concatenate((height[:-1,:],h),axis=0)
         cats = cats + 1
 
     h_min=h_mins[0]
     for h in h_mins[1:] :
-        h_min = np.concatenate((h_min[:-1,:],h),axis=0)    
+        h_min = np.concatenate((h_min[:-1,:],h),axis=0)
 
     h_max=h_maxs[0]
     for h in h_maxs[1:] :
-        h_max = np.concatenate((h_max[:-1,:],h),axis=0)    
+        h_max = np.concatenate((h_max[:-1,:],h),axis=0)
 
     h_std=h_stds[0]
     for h in h_stds[1:] :
-        h_std = np.concatenate((h_std[:-1,:],h),axis=0)    
+        h_std = np.concatenate((h_std[:-1,:],h),axis=0)
 
     h_x=h_xs[0]
     for h in h_xs[1:] :
-        h_x = np.concatenate((h_x[:-1,:],h),axis=0)    
+        h_x = np.concatenate((h_x[:-1,:],h),axis=0)
     h_y=h_ys[0]
     for h in h_ys[1:] :
-        h_y = np.concatenate((h_y[:-1,:],h),axis=0)    
-    #This is on supergrid. Topography is needed on tracer cells. 
-    #Sample every other point to get it. 
+        h_y = np.concatenate((h_y[:-1,:],h),axis=0)
+    #This is on supergrid. Topography is needed on tracer cells.
+    #Sample every other point to get it.
     #Drop the last x (corresponds to x0+360)  and y (why?)
     #Niki: Is this correct?
     #height= height[:-1:2,:-1:2]
